requirements.txt
asgiref==3.5.2
Django==4.1.1
django-cleanup==6.0.0
django-environ==0.9.0
django-tinymce==3.5.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.3
sqlparse==0.4.2
whitenoise==6.2.0
